refactor(screenshare): split 1564-line file into per-concern modules
No behavior change. Replaces commands/screenshare.rs with
commands/screenshare/ containing one file per concern:

  mod.rs              top-level re-exports, ScreenShareEvent,
                      LOCAL_PREVIEW_KEY, PREVIEW_MIN_INTERVAL,
                      fail_capture helper
  state.rs            HelperSession + ScreenShareState + impl
  codec.rs            pick_screenshare_codec, convert_to_i420,
                      pack_frame_bytes
  commands.rs         subscribe_screen_share_events/frames
  helper_subprocess.rs (cfg unix) spawn_and_accept_helper +
                      capture_helper_name + locate_capture_helper
  start_unix.rs       (cfg unix) enumerate, cancel_picker, start,
                      push_frame
  start_windows.rs    (cfg windows) enumerate, cancel_picker, start,
                      WindowsCapture{Flags,Handler}, push_frame_windows
  unsupported.rs      (cfg fallback) no-op stubs for non-tier-1 OS
  stop.rs             unified stop path
  remote_video.rs     on_remote_video_subscribed/unsubscribed/
                      on_room_disconnected

Public surface at pollis_core::commands::screenshare::* is unchanged.
Tests: 184 unit + 26 integration flows pass, same as baseline.
